move snowflake proxy to app module instead of orbotservice, way simpler, leverages existing translation. Add material.io icons for snowflake proxy on, starting+off
