+++ app/src/main/res/drawable/snowflake_off.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="960"
+    android:viewportHeight="960">
+  <path
+      android:pathData="M520,713v127q0,17 -11.5,28.5T480,880q-17,0 -28.5,-11.5T440,840v-126L338,815q-12,11 -28.5,10.5T282,814q-11,-11 -11,-28t11,-28l141,-142 -80,-80 -141,142q-11,11 -27.5,11.5T146,678q-11,-11 -11,-28t11,-28l100,-102L120,520q-17,0 -28.5,-11.5T80,480q0,-17 11.5,-28.5T120,440h127L59,252q-12,-12 -11.5,-28.5T60,195q12,-12 28.5,-12t28.5,12L796,875q12,12 12,28t-12,28q-12,12 -28.5,12T739,931L520,713ZM600,520h-45L440,405v-45L282,202q-11,-11 -11.5,-27.5T282,146q11,-11 28,-11t28,11l102,100v-126q0,-17 11.5,-28.5T480,80q17,0 28.5,11.5T520,120v126l102,-101q12,-11 28.5,-10.5T678,146q11,11 11,28t-11,28L520,360v80h80l158,-158q11,-11 27.5,-11.5T814,282q11,11 11,28t-11,28L714,440h126q17,0 28.5,11.5T880,480q0,17 -11.5,28.5T840,520L714,520l101,102q6,6 8.5,13.5t2.5,15q0,7.5 -3,14.5t-9,13q-11,11 -28,11t-28,-11L600,520Z"
+      android:fillColor="#e3e3e3"/>
+</vector>
