+++ app/src/main/res/drawable/snowflake_on.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="960"
+    android:viewportHeight="960">
+  <path
+      android:pathData="M440,714 L337,815q-11,11 -27.5,11T282,814q-12,-11 -12,-27.5t12,-28.5l158,-158v-80h-80L201,679q-11,11 -27.5,11T145,678q-11,-11 -11,-27.5t11,-27.5l101,-103L119,520q-17,0 -28,-11.5T80,480q0,-17 11.5,-28.5T120,440h126L145,338q-11,-11 -11,-27.5t12,-28.5q11,-11 27.5,-11t28.5,11l158,158h80v-80L281,202q-11,-11 -11,-27.5t12,-28.5q11,-11 27.5,-11t27.5,11l103,100v-126q0,-17 11.5,-28.5T480,80q17,0 28.5,11.5T520,120v126l102,-100q11,-11 27.5,-11t28.5,11q11,12 11,28.5T678,202L520,360v80h80l158,-158q11,-11 27.5,-11t28.5,12q11,11 11,27.5T814,338L714,440h126q17,0 28.5,11.5T880,480q0,17 -11.5,28.5T840,520L714,520l100,103q11,11 11,27.5T814,678q-12,12 -28.5,12T758,678L600,520h-80v80l158,159q11,11 11,27.5T677,815q-11,11 -27.5,11T622,815L520,714v127q0,17 -11.5,28T480,880q-17,0 -28.5,-11.5T440,840v-126Z"
+      android:fillColor="#e3e3e3"/>
+</vector>
