+++ app/src/main/res/drawable/snowflake_starting.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="960"
+    android:viewportHeight="960">
+  <path
+      android:pathData="M232,829q-12,11 -29,10.5T175,827q-35,-38 -57,-84t-22,-98q0,-50 14,-97t34,-92q15,-36 28,-72.5t13,-75.5q0,-35 -18,-65t-44,-54q-12,-11 -12,-27.5t11,-28.5q11,-12 27.5,-12.5T178,131q38,35 62.5,80t24.5,97q0,47 -13.5,91.5T219,487q-17,38 -30,76.5T176,644q0,37 16,69.5t42,59.5q11,12 10.5,28.5T232,829ZM392,829q-12,11 -29,10.5T335,827q-35,-38 -57,-84t-22,-98q0,-50 14,-97t34,-92q15,-36 28,-72.5t13,-75.5q0,-35 -18,-65t-44,-54q-12,-11 -12,-27.5t11,-28.5q11,-12 27.5,-12.5T338,131q38,35 62.5,80t24.5,97q0,47 -13.5,91T379,486q-17,38 -30,77t-13,81q0,37 16,69.5t42,59.5q11,12 10.5,28.5T392,829ZM480,840v-720q0,-17 11.5,-28.5T520,80q17,0 28.5,11.5T560,120v126l102,-101q11,-11 27.5,-11t28.5,12q11,11 11,28t-11,28L560,360v80h80l158,-158q11,-11 28,-11t28,11q11,11 11,28t-11,28L754,440h126q17,0 28.5,11.5T920,480q0,17 -11.5,28.5T880,520L754,520l101,102q11,11 11,27.5T854,678q-11,11 -28,11t-28,-11L640,520h-80v80l158,158q11,11 11,28t-11,28q-11,11 -28,11t-28,-11L560,714v126q0,17 -11.5,28.5T520,880q-17,0 -28.5,-11.5T480,840Z"
+      android:fillColor="#e3e3e3"/>
+</vector>
