+++ orbotservice/src/main/java/org/torproject/android/service/util/Prefs.kt
@@ -8,7 +8,6 @@ import android.preference.PreferenceManager
 import androidx.work.ExistingPeriodicWorkPolicy
 import androidx.work.PeriodicWorkRequest
 import androidx.work.WorkManager
-import org.torproject.android.service.circumvention.ResetSnowflakesServedWeeklyWorker
 import org.torproject.android.service.circumvention.Transport
 import java.util.Locale
 import java.util.concurrent.TimeUnit
